Code cleanup.

+++ src/stack.h
@@ -15,13 +15,11 @@
 
 #include <errno.h>
 
-
 #ifndef STACK_DEFAULT_SIZE
 #define STACK_DEFAULT_SIZE 64
 #endif
 
-//((void*)((int)stack->data+((stack->sp-1)*stack->item_size)))
-#define PTR_TO(stack, idx) ((void*) (((int) (stack)->data)+((stack)->item_size * (idx))) )
+#define PTR_TO(stack, idx) ((void*) (((stack)->data)+((stack)->item_size * (idx))) )
 #define STACK_LAST(stack) PTR_TO((stack), ((stack)->sp-1))
 
 typedef struct stack {
